+++ model_bakery/__about__.py
@@ -1,4 +1,4 @@
-__version__ = "1.8.0"
+__version__ = "1.9.0"
 __author__ = "berin"
 __email__ = "bernardoxhc@gmail.com"
 __url__ = "https://github.com/model-bakers/model_bakery"